New colors for the graph: - Temp and setting should have complementary colors, because the lines often overlap. This makes them hard to distinquish. - Beer setting and beer temperature now have the colors of the BrewPi logo. Green is made a bit darker for better visibility. Beer temperature has the color of the hops, beer setting of the raspberry. Fitting isn't it? The colors are almost exactly complementary. - The fridge color scheme is the same as the rest of the interface, also alsmost exactly complementary. - All 4 temperature are spread nicely over the color wheel for maximum contrast. - Room temp is a medium gray.
All lines now have thickness 1, because they are clearly visible with the new color scheme. Only the beer temp is 1.5 to emphasize it a bit. Thinner lines give better details without having to zoom in.
We might need to introduce a 'color blind' setting, because red and green are hard to distinguish for color blind people.
Display checkmark labels need updating to match the new color scheme.